Working but failling when connection is cut off

+++ include/application_layer.h
@@ -6,6 +6,7 @@
 
 #include "link_layer.h"
 #include <string.h>
+#include <stdlib.h>
 
 // Application layer main function.
 // Arguments:
@@ -19,7 +20,7 @@ void applicationLayer(const char *serialPort, const char *role, int baudRate,
                       int nTries, int timeout, const char *filename);
 
 int buildControlPacket(const long int file_size, const char *filename, unsigned char name_size, unsigned char* frame, int isStart);
-void buildDataPacket(short nDataBytes, unsigned char* packet, unsigned char* data);
+void buildDataPacket(int nDataBytes, unsigned char* packet, unsigned char* data);
 int getDataSize(unsigned char* data);
 
 #endif // _APPLICATION_LAYER_H_
